implementation de la video
